Moved main.c functions to board.c
