+++ stm32/Inc/board.h
@@ -1,0 +1,50 @@
+
+/**
+ * @brief Shared code for programs
+ *
+ *
+ * @author John Madden <jmadden173@pm.me>
+ * @date 2025-07-18
+ */
+
+#ifndef INC_BOARD_H_
+#define INC_BOARD_H_
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/**
+  * @brief System Clock Configuration
+  * @retval None
+  */
+void SystemClock_Config(void);
+
+/**
+* @brief  This function is executed in case of error occurrence.
+*
+* Turns on a solid status LED to give visual indicator to user. Enters an
+* infinite loop to halt program execution.
+*
+* @retval None
+*/
+void Error_Handler(void);
+
+
+
+#ifdef  USE_FULL_ASSERT
+/**
+  * @brief  Reports the name of the source file and the source line number
+  *         where the assert_param error has occurred.
+  * @param  file: pointer to the source file name
+  * @param  line: assert_param error line source number
+  * @retval None
+  */
+void assert_failed(uint8_t *file, uint32_t line);
+#endif /* USE_FULL_ASSERT */
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* INC_BOARD_H_ */
